Conflicts:
	Gemfile
	Gemfile.lock
	app/assets/javascripts/application.js.coffee
	app/assets/stylesheets/app/tables.css.scss
	app/controllers/administration/checkin/groups_controller.rb
	app/models/attendance_record.rb
	app/views/administration/checkin/groups/index.html.haml

+++ build/ubuntu/14.04/provision.sh
@@ -5,7 +5,7 @@ set -e
 sleep 10
 export DEBIAN_FRONTEND=noninteractive
 sudo apt-get update
-sudo -E apt-get install -y -q wget build-essential libcurl4-openssl-dev  libmysqlclient-dev nodejs git imagemagick mysql-server apache2
+sudo -E apt-get install -y -q wget build-essential libcurl4-openssl-dev libmysqlclient-dev nodejs git imagemagick mysql-server apache2 libapache2-mod-xsendfile
 
 sudo apt-get install -y software-properties-common
 sudo apt-add-repository -y ppa:brightbox/ruby-ng
@@ -20,7 +20,8 @@ sudo chmod 600 /etc/apt/sources.list.d/passenger.list
 sudo apt-get update
 sudo apt-get install -y libapache2-mod-passenger
 sudo a2enmod passenger
-sudo sed -i "s/DocumentRoot.*/DocumentRoot \/var\/www\/onebody\/public/" /etc/apache2/sites-available/000-default.conf
+sudo a2enmod xsendfile
+sudo sed -i "s/DocumentRoot.*/DocumentRoot \/var\/www\/onebody\/public\n\nXSendFile On\nXSendFilePath \/var\/www\/onebody\/public\/system/" /etc/apache2/sites-available/000-default.conf
 sudo service apache2 restart
 
 if [[ `grep RAILS_ENV .bashrc` == "" ]]; then
@@ -33,10 +34,18 @@ sudo chown -R $USER /var/www/onebody
 
 cd /var/www/onebody
 git checkout stable
+
+# brightbox may have a slightly newer version than we're pinned to -- that's ok
+sed -i '/ruby-version/d' Gemfile
+
 bundle install --deployment
 
 mysql -uroot -e "create database if not exists onebody default character set utf8 default collate utf8_general_ci; grant all on onebody.* to onebody@localhost identified by 'onebody';"
-cp config/database.yml{.example,}
+if [[ -e config/database.yml.mysql-example ]]; then
+  cp config/database.yml{.mysql-example,}
+else
+  cp config/database.yml{.example,}
+fi
 RAILS_ENV=production bundle exec rake db:migrate db:seed
 
 cp config/secrets.yml{.example,}
@@ -46,7 +55,7 @@ sed -i "s/SOMETHING_RANDOM_HERE/$secret/" config/secrets.yml
 export DEBIAN_FRONTEND=noninteractive
 sudo -E apt-get install -y -q postfix courier-pop
 sudo sed -i "s/mydestination.*/mydestination = localhost, your-domain-goes-here.com/" /etc/postfix/main.cf
-if [[ `grep Mialdir /etc/postfix/main.cf` == "" ]]; then
+if [[ `grep Maildir /etc/postfix/main.cf` == "" ]]; then
   echo -e "home_mailbox = Maildir/\nsmtp_discard_ehlo_keywords=pipelining,discard\nmessage_size_limit = 25600000\nlocal_recipient_maps =\nluser_relay = onebodymail" | sudo tee -a /etc/postfix/main.cf
 fi
 sudo postfix reload
@@ -56,7 +65,7 @@ if [[ ! -e /home/onebodymail ]]; then
   sudo adduser --gecos "" --disabled-password --home=/home/onebodymail onebodymail
   email_password=`bundle exec rake secret | head -c 16`
   echo -e "$email_password\n$email_password" | sudo passwd onebodymail
-  echo -e "production:\n  pop:\n    host: localhost\n    username: onebodymail\n    password: $email_password\n  smtp:\n    address: localhost\n    domain: example.com" > config/email.yml
+  echo -e "production:\n  pop:\n    host: localhost\n    username: onebodymail\n    password: $email_password\n  smtp:\n    address: localhost\n    domain: example.com\n    enable_starttls_auto: false" > config/email.yml
 fi
 
 bundle exec whenever -w
